+++ src/structures/Watchdog.ts
@@ -1,14 +1,17 @@
 import flatMap from "array.prototype.flatmap";
-import Eris, { Client } from "eris";
+import { addMinutes, formatDistanceToNow } from "date-fns";
+import { format } from "date-fns-tz";
+import Eris, { Client, Embed } from "eris";
 import LRU from "lru-cache";
+import fetch from "node-fetch";
 import path, { resolve as res } from "path";
 import pluralize from "pluralize";
 import { GatewayServer, SlashCreator } from "slash-create";
 import { walk } from "walk";
 import LogHandler from "../handlers/logHandler";
 import { mentionRole, sendWebhookMessage } from "../services/Discord";
-import { CreateAccount, Login } from "../services/PlayFab";
-import config, { Role } from "../structures/Config";
+import { CreateAccount, getServerInfo, Login } from "../services/PlayFab";
+import config, { Role, Server } from "../structures/Config";
 import logger from "../utils/logger";
 import MordhauAPI from "../utils/MordhauAPI";
 import { outputPlayerIDs } from "../utils/PlayerID";
@@ -17,6 +20,7 @@ import AntiSlur from "./AutoMod";
 import AutoUpdater from "./AutoUpdater";
 import BaseRCONCommand from "./BaseRCONCommands";
 import Database from "./Database";
+import DiscordEmbed from "./DiscordEmbed";
 import Rcon from "./Rcon";
 
 interface Iids {
@@ -26,7 +30,7 @@ interface Iids {
 }
 
 export default class Watchdog {
-    private client: Client;
+    public client: Client;
     private token: string;
     public startTime: number = Date.now();
     public autoUpdater: AutoUpdater;
@@ -79,8 +83,231 @@ export default class Watchdog {
             };
         }
     >;
-
     public logHandler: LogHandler;
+
+    async sendOrUpdateStatuses(server: { name: string; rcon: Rcon }) {
+        try {
+            const configServer = config
+                .get("servers")
+                .find((s) => s.name === server.name);
+            const channelID = configServer.rcon.status.channel;
+            if (!channelID) {
+                return logger.debug(
+                    "Server Status",
+                    `Skipping ${server.name} status message`
+                );
+            }
+
+            logger.debug(
+                "Server Status",
+                `Updating ${
+                    server.name
+                } status message (Next update: ${formatDistanceToNow(
+                    addMinutes(
+                        new Date(),
+                        configServer.rcon.status.updateInterval
+                    ),
+                    { addSuffix: true }
+                )})`
+            );
+
+            const { online, currentMap, gamemode, name } =
+                await server.rcon.getServerInfo();
+            const players = await server.rcon.getIngamePlayers();
+            // const players = await Promise.all(
+            //     (
+            //         await server.rcon.getIngamePlayers()
+            //     ).map(
+            //         async (p) =>
+            //             this.cachedPlayers.get(p.id) ||
+            //             (await LookupPlayer(p.id))
+            //     )
+            // );
+            const serverInfo = await getServerInfo({
+                name,
+                host: server.rcon.options.host,
+                port: server.rcon.options.port,
+            });
+            const adress = `${server.rcon.options.host}:${serverInfo.ServerPort}`;
+            const country = await (
+                await fetch(
+                    `https://ipinfo.io/${server.rcon.options.host}/country`
+                )
+            ).text();
+            const maxPlayerCount =
+                (serverInfo
+                    ? parseInt(serverInfo.Tags.MaxPlayers)
+                    : server.rcon.maxPlayerCount) || 0;
+            const currentPlayerCount = serverInfo
+                ? serverInfo.Tags.Players.split(",").filter((p) => p.length)
+                      .length
+                : 0;
+            const passwordProtected = serverInfo
+                ? serverInfo.Tags.IsPasswordProtected === "true"
+                : false;
+
+            let message: Eris.Message<Eris.TextableChannel>;
+
+            try {
+                message = server.rcon.statusMessageID
+                    ? await this.client.getMessage(
+                          channelID,
+                          server.rcon.statusMessageID
+                      )
+                    : (await this.client.getMessages(channelID))?.find((m) => {
+                          const footer = m.embeds[0].footer.text.split(" | ");
+                          return m.author.id === this.client.user.id &&
+                              !configServer.rcon.status.hideIPPort
+                              ? m.embeds[0]?.fields.some(
+                                    (f) =>
+                                        f.name === "Address:Port" &&
+                                        f.value === `\`${adress}\``
+                                )
+                              : parseInt(footer[footer.length - 1]) - 1 ===
+                                    config
+                                        .get("servers")
+                                        .findIndex(
+                                            (s: Server) =>
+                                                s.rcon.host ===
+                                                    configServer.rcon.host &&
+                                                s.rcon.port ===
+                                                    configServer.rcon.port
+                                        );
+                      });
+            } catch {}
+
+            async function generateStatusMessage(baseEmbed?: Embed) {
+                const embed = new DiscordEmbed();
+                const date = new Date();
+
+                embed
+                    .setTitle(
+                        name
+                            ? `${passwordProtected ? ":lock: " : ""}\`${name}\``
+                            : baseEmbed?.title || "Unknown"
+                    )
+                    .addField(
+                        "Status",
+                        online
+                            ? `:green_circle: **Online**`
+                            : `:red_circle: **Offline**`,
+                        true
+                    );
+
+                if (!configServer.rcon.status.hideIPPort)
+                    embed
+                        .setDescription(`Connect: steam://connect/${adress}`)
+                        .addField("Address:Port", `\`${adress}\``, true);
+
+                embed
+                    .addField(
+                        "Location",
+                        country
+                            ? `:flag_${country
+                                  .toLowerCase()
+                                  .trim()}: ${country}`
+                            : baseEmbed?.fields?.find(
+                                  (f) => f.name === "Location"
+                              )?.value || ":united_nations: Unknown",
+                        true
+                    )
+                    .addField(
+                        "Gamemode",
+                        !gamemode
+                            ? baseEmbed?.fields?.find(
+                                  (f) => f.name === "Gamemode"
+                              ).value || "Unknown"
+                            : `${gamemode || "Unknown"}`,
+                        true
+                    )
+                    .addField(
+                        "Current Map",
+                        !currentMap
+                            ? baseEmbed?.fields?.find(
+                                  (f) => f.name === "Current Map"
+                              ).value || "Unknown"
+                            : `${currentMap || "Unknown"}`,
+                        true
+                    )
+                    .addField(
+                        `Players${
+                            configServer.rcon.status.showPlayerList
+                                ? ` ${currentPlayerCount}/${maxPlayerCount}`
+                                : ""
+                        }`,
+                        !configServer.rcon.status.showPlayerList
+                            ? `${currentPlayerCount}/${maxPlayerCount}`
+                            : `\`\`\`${
+                                  players
+                                      .map((p) => `${p.id} - ${p.name}`)
+                                      .join("\n") || "No players online"
+                              }\`\`\``,
+                        !configServer.rcon.status.showPlayerList ? true : false
+                    )
+                    .setFooter(
+                        `Mordhau RCON | Last Update: ${format(
+                            addMinutes(date, date.getTimezoneOffset()),
+                            "yyyy-MM-dd HH:mm:ss"
+                        )} UTC${
+                            configServer.rcon.status.hideIPPort
+                                ? ` | ${
+                                      config
+                                          .get("servers")
+                                          .findIndex(
+                                              (s: Server) =>
+                                                  s.rcon.host ===
+                                                      configServer.rcon.host &&
+                                                  s.rcon.port ===
+                                                      configServer.rcon.port
+                                          ) + 1
+                                  }`
+                                : ""
+                        }`
+                    );
+
+                return embed;
+            }
+
+            if (!message) {
+                const embed = await generateStatusMessage();
+
+                const m = await this.client.createMessage(channelID, {
+                    embed: embed.getEmbed(),
+                });
+                server.rcon.statusMessageID = m.id;
+            } else {
+                server.rcon.statusMessageID = message.id;
+
+                const messageEmbed = message.embeds[0];
+                const baseEmbed = new DiscordEmbed()
+                    .setTitle(messageEmbed.title)
+                    .setDescription(messageEmbed.description);
+
+                for (let i = 0; i < messageEmbed.fields.length; i++) {
+                    const field = messageEmbed.fields[i];
+                    baseEmbed.addField(field.name, field.value, field.inline);
+                }
+
+                const embed = await generateStatusMessage(messageEmbed);
+
+                this.client.editMessage(
+                    channelID,
+                    server.rcon.statusMessageID,
+                    {
+                        embed: embed.getEmbed(),
+                    }
+                );
+            }
+        } catch (error) {
+            logger.error(
+                "Server Status",
+                `Error occurred while updating ${server.name} status (Error: ${
+                    error.message || error
+                })`
+            );
+        }
+    }
+
     public rcon = {
         getServersInfo: async () => {
             const results: {
@@ -987,7 +1214,8 @@ export default class Watchdog {
                                     | "permanent"
                                     | "automod"
                                     | "killstreak"
-                                    | "adminCalls",
+                                    | "adminCalls"
+                                    | "warns",
                                 {
                                     id: webhook.id,
                                     token: webhook.token,
@@ -1036,7 +1264,8 @@ export default class Watchdog {
                                         | "permanent"
                                         | "automod"
                                         | "killstreak"
-                                        | "adminCalls",
+                                        | "adminCalls"
+                                        | "warns",
                                     {
                                         id: newWebhook.id,
                                         token: newWebhook.token,
@@ -1111,7 +1340,14 @@ export default class Watchdog {
         await this.loadRCONCommands();
 
         for (const [name, server] of this.servers) {
+            const s = config.get("servers").find((s) => s.name === name);
+
             server.rcon.initialize();
+
+            setInterval(
+                () => this.sendOrUpdateStatuses(server),
+                s.rcon.status.updateInterval * 60 * 1000
+            );
         }
 
         logger.info("Bot", "Client initialized - running client.");